+++ app/src/global.js
@@ -34,7 +34,10 @@ ytd-compact-video-renderer,
 ytd-grid-video-renderer,
 ytd-playlist-video-renderer,
 ytd-playlist-panel-video-renderer,
-yt-lockup-view-model`,
+yt-lockup-view-model,
+ytm-compact-video-renderer,
+ytm-rich-item-renderer,
+ytm-video-with-context-renderer`,
   ALL_ARRAYS_SHORTS_SELECTOR: `div.style-scope.ytd-rich-item-renderer,
 ytm-shorts-lockup-view-model`,
   cacheSessionStorageKey: "[YoutubeAntiTranslate]cache",
@@ -109,11 +112,12 @@ ytm-shorts-lockup-view-model`,
    * @returns {string}
    */
   getPlayerSelector: function () {
-    this.logDebug(`getPlayerSelector called`);
+    if (window.location.hostname === "m.youtube.com") {
+      return "#player-container-id";
+    }
     const selector = window.location.pathname.startsWith("/shorts")
       ? "#shorts-player"
       : "ytd-player .html5-video-player";
-    this.logDebug(`getPlayerSelector returning: ${selector}`);
     return selector;
   },
 
@@ -121,9 +125,7 @@ ytm-shorts-lockup-view-model`,
    * @returns {string}
    */
   getBrowserOrChrome: function () {
-    this.logDebug(`getBrowserOrChrome called`);
     const result = typeof browser !== "undefined" ? browser : chrome;
-    this.logDebug(`getBrowserOrChrome returning browser type`);
     return result;
   },
 
@@ -131,12 +133,16 @@ ytm-shorts-lockup-view-model`,
    * @returns {bool}
    */
   isFirefoxBasedBrowser: function () {
-    this.logDebug(`isFirefoxBasedBrowser called`);
     const result =
       typeof browser !== "undefined" &&
       typeof browser.runtime !== "undefined" &&
       typeof browser.runtime.getBrowserInfo === "function";
-    this.logDebug(`isFirefoxBasedBrowser returning: ${result}`);
+    return result;
+  },
+
+  // Detects if we are currently on the mobile YouTube site (m.youtube.com)
+  isMobile: function () {
+    const result = window.location.hostname === "m.youtube.com";
     return result;
   },
 
@@ -977,5 +983,11 @@ ytm-shorts-lockup-view-model`,
       `detectSupportedLanguage: all attempts exhausted, returning null`,
     );
     return null;
+  },
+  getSettings: function () {
+    const element = document.querySelector(
+      'script[type="module"][data-ytantitranslatesettings]',
+    );
+    return JSON.parse(element?.dataset?.ytantitranslatesettings ?? "{}");
   },
 };